requirements.txt
autopep8==1.5.5
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.3.4
networkx==2.5
numpy==1.19.5
Pillow==8.1.0
pkg-resources==0.0.0
pycocotools==2.0.0
pycodestyle==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.4
six==1.15.0
tifffile==2020.9.3
toml==0.10.2
